--- a/app/global-error.tsx
+++ b/app/global-error.tsx
@@ -3,7 +3,7 @@
 import Link from 'next/link';
 import { Button } from '@/components/ui/button';
 import { ArrowLeft, RefreshCw } from 'lucide-react';
-import * as Sentry from '@sentry/nextjs';
+import { Sentry } from '@/lib/sentry';
 import { useEffect } from 'react';
 
 export default function GlobalError({