--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -50,12 +50,12 @@ services:
       context: .
       target: devrunner
     depends_on:
-      database:
-        condition: service_healthy
+      - database
     ports:
       - "${DEV_APP_PORT:-3000}:3000"
     volumes:
       - ./:/app
+      - ./lib/sentry-dev.ts:/app/lib/sentry.ts
       # - ./dev_modules/app/node_modules:/app/node_modules
       - ./dev_modules/app/.next:/app/.next
       - ./dev_modules/pnpm_store:/app/.pnpm-store