--- a/instrumentation-client.ts
+++ b/instrumentation-client.ts
@@ -2,7 +2,7 @@
 // The added config here will be used whenever a users loads a page in their browser.
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
-import * as Sentry from '@sentry/nextjs';
+import { Sentry } from '@/lib/sentry';
 
 if (process.env.SENTRY_DSN) {
   Sentry.init({