--- a/lib/sentry.ts
+++ b/lib/sentry.ts
@@ -1,0 +1,6 @@
+// this is the default/production sentry object
+// it adds about 2k modules in next dev build and hurts
+// the developer experience.
+import * as Sentry from '@sentry/nextjs';
+
+export { Sentry };