--- a/sentry.edge.config.ts
+++ b/sentry.edge.config.ts
@@ -3,7 +3,7 @@
 // Note that this config is unrelated to the Vercel Edge Runtime and is also required when running locally.
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
-import * as Sentry from '@sentry/nextjs';
+import { Sentry } from '@/lib/sentry';
 
 if (process.env.SENTRY_DSN) {
   Sentry.init({