--- a/sentry.server.config.ts
+++ b/sentry.server.config.ts
@@ -2,7 +2,7 @@
 // The config you add here will be used whenever the server handles a request.
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
-import * as Sentry from '@sentry/nextjs';
+import { Sentry } from '@/lib/sentry';
 
 if (process.env.SENTRY_DSN) {
   Sentry.init({