- in dev mode, avoid loading any sentry modules which speeds up the build-time significantly

- moves all @sentry/nextjs imports to @/lib/sentry
- magic with docker-compose to override @/lib/sentry.ts with a mocked @/lib/sentry-dev.ts
- magic with docker-compose to start app even if db is not ready yet (saves ~5s)
- only load withSentryConfig in next.config if in production

reduces
 - initial build/load from 961 modules to 24 modules, 4.5s to 1.5s
 - initial load of '/' from 3912 modules to 1522 modules, 10.5s to 6.4s

(numbers with all wagmi/ethers/silk disabled as well)